# apps/api/openapi-v2.yml
openapi: 3.0.3
info:
  title: Edge-Cloud Orchestrator API
  description: Production API documentation for the Edge-Cloud Orchestrator Control Plane
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    def-0:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error description
            requestId:
              type: string
              description: Unique request ID
            timestamp:
              type: string
              format: date-time
              description: ISO 8601 timestamp
            details:
              description: Optional error details
            stack:
              type: string
              description: Error stack trace (non-production only)
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
      title: ErrorSchema
...
paths:
...
  /v2/agents/certificates/sign:
    post:
      summary: Bootstrap agent certificate and register node
      tags:
        - agents
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - csr
                - bootstrapToken
                - nodeName
                - region
              properties:
                csr:
                  type: string
                  description: PEM encoded Certificate Signing Request
                bootstrapToken:
                  type: string
                  description: One-time bootstrap token
                nodeName:
                  type: string
                region:
                  type: string
                ipAddress:
                  type: string
                port:
                  type: number
                cpuCores:
                  type: number
                memoryGB:
                  type: number
                storageGB:
                  type: number
                hardwareId:
                  type: string
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificate:
                    type: string
                  caCertificate:
                    type: string
                  nodeId:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /v2/agents/heartbeat:
    post:
      summary: Agent heartbeat
      tags:
        - agents
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                timestamp:
                  type: number
                status:
                  type: string
                metrics:
                  type: object
                  properties:
                    cpu_usage:
                      type: number
                    memory_usage:
                      type: number
      responses:
        "200":
          description: Default Response
  /v2/agents/tasks/pending:
    get:
      summary: Poll for pending tasks
      tags:
        - agents
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  runtime:
                    type: string
                  image:
                    type: string
                  wasm_artifact_id:
                    type: string
                    nullable: true
                  input:
                    type: object
                    additionalProperties: true
                  memory_limit_mb:
                    type: number
                  cpu_fuel:
                    type: number
                    nullable: true
                  timeout_seconds:
                    type: number
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /v2/agents/tasks/{taskId}/result:
    post:
      summary: Report task result
      tags:
        - agents
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - status
                - exit_code
              properties:
                status:
                  type: string
                exit_code:
                  type: number
                stdout:
                  type: string
                stderr:
                  type: string
                duration_ms:
                  type: number
                error:
                  type: string
                  nullable: true
                metrics:
                  type: object
                  properties:
                    cpu_usage_avg:
                      type: number
                    memory_usage_max:
                      type: number
                    network_ingress_bytes:
                      type: number
                    network_egress_bytes:
                      type: number
        required: true
      parameters:
        - schema:
            type: string
          in: path
          name: taskId
          required: true
      responses:
        "200":
          description: Default Response
  /v2/agents/tasks/{taskId}/status:
    post:
      summary: Update task execution status
      tags:
        - agents
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
        required: true
      parameters:
        - schema:
            type: string
          in: path
          name: taskId
          required: true
      responses:
        "200":
          description: Default Response
...